Tweak milestone style slightly and make gantt bars not wrap, fixes bug
2004-10-06  Richard Hult  <richard@imendio.com>

	* data/stylesheets/html1_css.xsl:
	* data/stylesheets/html1_gantt.xsl:
	* data/stylesheets/mrproject2html.xsl: Tweak milestone style
	slightly and make gantt bars not wrap, fixes bug #154572.

	* libplanner/mrp-parser.c (mrp_parser_to_xml): Fix compilation
	problem on Darwin.

	* configure.in: Clean up a bit.

+++ libplanner/mrp-parser.c
@@ -1135,7 +1135,7 @@ mrp_parser_to_xml (MrpStorageMrproject  *module,
 {
 	xmlDocPtr  doc;
 	xmlChar   *buf;
-	gsize      len;
+	int        len;
 	
 	g_return_val_if_fail (MRP_IS_STORAGE_MRPROJECT (module), FALSE);
